Move JoinType to datafusion_common (#6572)
